backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.0",
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-elasticsearch-dsl>=8.2",
    "django-prometheus==2.5.0.dev12",
    "django-redis>=6.0.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "elasticsearch>=8.19.2",
    "djangorestframework-simplejwt>=5.5.1",
    "factory-boy>=3.3.3",
    "fakeredis>=2.34.0",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "channels>=4.1",
    "channels-redis>=4.2",
    "daphne>=4.1",
    "pytest-asyncio>=0.23",
    "sendgrid>=6.11",
    "firebase-admin>=6.5",
    "prometheus-client>=0.24.1",
    "sentry-sdk[celery,django,redis]>=2.55.0",
    "drf-spectacular>=0.29.0",
    "django-ratelimit>=4.1.0",
    "locust>=2.43.3",
    "bandit>=1.9.4",
    "safety>=3.7.0",
    "gunicorn>=25.1.0",
]

[project.optional-dependencies]
dev = ["locust", "bandit", "safety"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"
markers = [
    "unit: fast, isolated tests with no DB/network",
    "integration: tests that hit the real DB",
    "e2e: full API request/response tests",
    "slow: tests that take more than 1s",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/factories.py",
]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = ["S101", "S106"]   # allow assert and hardcoded passwords in tests
"*/factories.py" = ["S106"]
"config/settings/*.py" = ["F403", "F405"]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "black>=26.3.1",
    "locust>=2.43.3",
    "mypy>=1.19.1",
    "ruff>=0.15.7",
    "safety>=3.7.0",
    "types-requests>=2.32.4.20260324",
]
